requirements.txt
aiogram~=2.18
aioredis<2.0
captcha>=0.4
environs~=9.0
PyYAML>=6.0
motor>=3.1.1
pymongo>=4.3.3